Address PR review comments: extract useHoldMenuSubmit, re-add iOS InteractionManager guard
